Added preallocate_file() function which acts as a passthrough to SYS_fallocate, fallocate, or posix_fallocate depending on the target configuration.
As I mentioned to Joey, my machine has none of these, and I haven't had a chance to test on a SEAS machine, so I can't guarantee that it works on each possible config. However, it *should* work, and at least serves as a starting point so we can start placing calls to this function throughout the code.